+++ egui/src/lib.rs
@@ -27,7 +27,7 @@
 //!     egui_ctx.begin_frame(raw_input);
 //!     my_app.ui(&egui_ctx); // add panels, windows and widgets to `egui_ctx` here
 //!     let (output, paint_commands) = egui_ctx.end_frame();
-//!     let paint_jobs = egui_ctx.tesselate(paint_commands); // create triangles to paint
+//!     let paint_jobs = egui_ctx.tessellate(paint_commands); // create triangles to paint
 //!     my_integration.paint(paint_jobs);
 //!     my_integration.set_cursor_icon(output.cursor_icon);
 //!     // Also see `egui::Output` for more
@@ -143,7 +143,7 @@ fn test_egui_e2e() {
         ctx.begin_frame(raw_input.clone());
         demo_windows.ui(&ctx, &Default::default(), &mut None, |_ui| {});
         let (_output, paint_commands) = ctx.end_frame();
-        let paint_jobs = ctx.tesselate(paint_commands);
+        let paint_jobs = ctx.tessellate(paint_commands);
         assert!(!paint_jobs.is_empty());
     }
 }